Fix bugs (ws connection and servo time overflow)

+++ arduino/rfid_servo/status.h
@@ -9,7 +9,7 @@ struct Status {
     // access control
     bool authorized;
     bool unlocked;
-    unsigned int unlock_time;
+    unsigned long unlock_time;
 };
 
 extern Status status;
